% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ganancia.R
\name{Ganancia}
\alias{Ganancia}
\title{Genera la ganancia total}
\usage{
Ganancia(Ganancia, carne)
}
\arguments{
\item{Ganancia}{(data.frame) Datos respuesta}

\item{carne}{(data.frame) Datos que explican la variable respuesta}
}
\value{
Un solo dato, con la ganancia en pesos de un año
}
\description{
Realiza una suma de la columna venta
}
\examples{
\dontrun{
#Directorio del trabajo
ruta<- "C:/Users/dmake/OneDrive/Escritorio/DatosProg"

#------------------------------------------------------
#Ejemplo
datos1<- openxlsx::read.xlsx("parametrosR.xlsx")
#cargar libreria
library(ProgAnual)
programacion(par=datos1$parametros, x=datos1$x )
}
}
