% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prog_Anual.R
\name{programacion}
\alias{programacion}
\title{Programación Anual De Una Granja Cunicola}
\usage{
programacion(par, x)
}
\arguments{
\item{par}{(cuadro) nombre de las variables}

\item{x}{(cuadro) datos numericos de la variable explicativa}
}
\value{
un data.frame (calendarizacion) con valores de produccion y reproduccion de manera quincenal en un año
}
\description{
Realiza una tabla con datos de hembras gestantes, lactantes; gazapos "1", gazapos "2"; conejos en fase de engorda; conejos finalizados; kg de carne obtenidos; dinero generado por la venta de esos kg de carne. Todo esto obtenido de la utilizacion de diferentes parametros reproductivos y productivos dados de una manera quincenal
}
\examples{
\dontrun{
#Directorio del trabajo
ruta<- "C:/Users/dmake/OneDrive/Escritorio/DatosProg"

#------------------------------------------------------
#Ejemplo1
datos1<- openxlsx::read.xlsx("parametrosR.xlsx")
#cargar libreria
library(ProgAnual)
programacion(par=datos1$parametros, x=datos1$x )
}
}
